remove useless comment

+++ sources/cl/calc_object.cl
@@ -49,7 +49,7 @@ float3			ray_cone_norm(global t_obj *obj, float3 ray_pos, float3 ray_dir, short 
 	offset = ray_pos - obj->pos;
 	a = dot(ray_dir.xz, ray_dir.xz) - dot(ray_dir.y, ray_dir.y);
 	b = 2 * (dot(ray_dir.xz, offset.xz) - dot(ray_dir.y, ray_dir.y)) -
-		(float)obj->radius * obj->radius;
+		obj->radius * obj->radius;
 	c = dot(offset.xz, offset.xz) - dot(offset.y, offset.y);
 	if ((delta = calc_delta(a, b, c)) >= 0)
 		*ok = -1;
@@ -71,7 +71,7 @@ float3			ray_cylinder_norm(global t_obj *obj, float3 ray_pos, float3 ray_dir, sh
 	ray_dir.y = 0;
 	a = dot(ray_dir, ray_dir);
 	b = 2.0f * dot(ray_dir, offset);
-	c = dot(offset, offset) - (float)obj->radius * obj->radius;
+	c = dot(offset, offset) - obj->radius * obj->radius;
 	if ((delta = calc_delta(a, b, c)) < 0.0f)
 		*ok = -1;
 	return (norm(delta, ray_pos, rdir));
@@ -88,7 +88,7 @@ float3			ray_sphere_norm(global t_obj *obj, float3 ray_pos, float3 ray_dir, shor
 	offset = ray_pos - obj->pos;
 	a = dot(ray_dir, ray_dir);
 	b = 2.0f * dot(ray_dir, offset);
-	c = dot(offset, offset) - (float)(obj->radius * obj->radius);
+	c = dot(offset, offset) - (obj->radius * obj->radius);
 	if ((delta = calc_delta(a, b, c)) < 0.0f)
 		*ok = -1;
 	return (norm(delta, ray_pos, ray_dir));
